// src/app/pages/models/user.model.ts
export class User {
  $key: string;
  'user-id': string;
  username: string;
  password: string;
  building: string;
  emailid: string;
  extrafeatures: string;
  flat: string;
  housetype: string;
  landline: string;
  mobile: string;
  name: string;
  parking: boolean;
  salutationid: number;
  slotno: string;
  stickerno: string;
  twofourwheel: string;
  vmodel: string;
  selectionFeature: any;
  carpetarea: number;
}